inverse control for addAuthor and deleteAuthor functions
