+++ src/components/CreateCourse/components/Authors/Authors.tsx
@@ -1,10 +1,32 @@
 import React, { FC, useContext } from 'react';
 import AuthorItem from '../AuthorItem/AuthorItem';
 import { CoursesContext } from 'src/context/CourseContext';
+import { Author } from 'src/types/common/types';
 
 export const Authors: FC = () => {
-	const { authors, chosenAuthors } = useContext(CoursesContext);
+	const { authors, chosenAuthors, setChosenAuthors } =
+		useContext(CoursesContext);
 
+	const choseAuthor = (author: Author) => {
+		// Check if the author is already in the chosenAuthors array
+		const isAuthorAlreadyChosen = chosenAuthors.some(
+			(chosenAuthor) => chosenAuthor.id === author.id
+		);
+
+		if (!isAuthorAlreadyChosen) {
+			// If the author is not already chosen, add it to the chosenAuthors array
+			const updatedChosenAuthors = [...chosenAuthors, author];
+			setChosenAuthors(updatedChosenAuthors);
+		}
+	};
+
+	const deleteAuthor = (author: Author) => {
+		// Filter out the selected author from chosenAuthors array
+		const updatedChosenAuthors = chosenAuthors.filter(
+			(chosenAuthor) => chosenAuthor.id !== author.id
+		);
+		setChosenAuthors(updatedChosenAuthors);
+	};
 	return (
 		<div className='flex flex-col sm:justify-center basis-full sm:basis-6/12 text-center sm:p-6'>
 			<div>
@@ -15,6 +37,8 @@ export const Authors: FC = () => {
 							isInChosenList={false}
 							key={author.id}
 							author={author}
+							handleDeletion={deleteAuthor}
+							addAuthor={choseAuthor}
 						/>
 					))}
 				</div>
@@ -28,6 +52,8 @@ export const Authors: FC = () => {
 								isInChosenList={true}
 								key={author.id}
 								author={author}
+								handleDeletion={deleteAuthor}
+								addAuthor={choseAuthor}
 							/>
 						))}
 				</div>
